refactor(data_operations): move shared arithmetic package under row_preserving
The shared arithmetic building blocks (base, per-backend mixins,
numeric-source descriptors) sat at data_operations/arithmetic/, a sibling
of row_preserving/ as if it were a top-level operation category. Its only
consumers (point_arithmetic, scalar_arithmetic) live under row_preserving/,
so the shared code belongs beside them, not above them.

Move it to data_operations/row_preserving/arithmetic/ and update all
imports. It still ships in the mloda-community-data-operations base wheel
that both arithmetic families already depend on, so the published
dependency graph is unchanged (no leaf-to-leaf wheel edge).
