+++ mloda/community/feature_groups/data_operations/row_preserving/arithmetic/duckdb_numeric_source.py
