+++ mloda/community/feature_groups/data_operations/row_preserving/arithmetic/pandas_mixin.py
@@ -19,12 +19,12 @@ import pandas as pd
 from mloda.provider import ComputeFramework
 from mloda_plugins.compute_framework.base_implementations.pandas.dataframe import PandasDataFrame
 
-from mloda.community.feature_groups.data_operations.arithmetic.pandas_numeric_source import (
+from mloda.community.feature_groups.data_operations.row_preserving.arithmetic.pandas_numeric_source import (
     pandas_non_numeric_descriptor,
 )
 
 if TYPE_CHECKING:
-    from mloda.community.feature_groups.data_operations.arithmetic.base import ArithmeticFeatureGroupBase
+    from mloda.community.feature_groups.data_operations.row_preserving.arithmetic.base import ArithmeticFeatureGroupBase
 
     _ArithmeticMixinBase = ArithmeticFeatureGroupBase
 else:
