+++ mloda/community/feature_groups/data_operations/row_preserving/arithmetic/polars_mixin.py
@@ -19,12 +19,12 @@ import polars as pl
 from mloda.provider import ComputeFramework
 from mloda_plugins.compute_framework.base_implementations.polars.lazy_dataframe import PolarsLazyDataFrame
 
-from mloda.community.feature_groups.data_operations.arithmetic.polars_numeric_source import (
+from mloda.community.feature_groups.data_operations.row_preserving.arithmetic.polars_numeric_source import (
     polars_non_numeric_descriptor,
 )
 
 if TYPE_CHECKING:
-    from mloda.community.feature_groups.data_operations.arithmetic.base import ArithmeticFeatureGroupBase
+    from mloda.community.feature_groups.data_operations.row_preserving.arithmetic.base import ArithmeticFeatureGroupBase
 
     _ArithmeticMixinBase = ArithmeticFeatureGroupBase
 else:
