+++ mloda/community/feature_groups/data_operations/row_preserving/arithmetic/sqlite_mixin.py
@@ -19,13 +19,13 @@ from mloda.provider import ComputeFramework
 from mloda_plugins.compute_framework.base_implementations.sqlite.sqlite_framework import SqliteFramework
 from mloda_plugins.compute_framework.base_implementations.sqlite.sqlite_relation import SqliteRelation
 
-from mloda.community.feature_groups.data_operations.arithmetic.base import SQL_ARITHMETIC_OPS
-from mloda.community.feature_groups.data_operations.arithmetic.sqlite_numeric_source import (
+from mloda.community.feature_groups.data_operations.row_preserving.arithmetic.base import SQL_ARITHMETIC_OPS
+from mloda.community.feature_groups.data_operations.row_preserving.arithmetic.sqlite_numeric_source import (
     sqlite_non_numeric_descriptor,
 )
 
 if TYPE_CHECKING:
-    from mloda.community.feature_groups.data_operations.arithmetic.base import ArithmeticFeatureGroupBase
+    from mloda.community.feature_groups.data_operations.row_preserving.arithmetic.base import ArithmeticFeatureGroupBase
 
     _ArithmeticMixinBase = ArithmeticFeatureGroupBase
 else:
