+++ mloda/community/feature_groups/data_operations/row_preserving/arithmetic/sqlite_numeric_source.py
