+++ mloda/community/pyproject.toml
@@ -20,4 +20,4 @@ Repository = "https://github.com/mloda-ai/mloda-registry"
 
 [tool.setuptools]
 package-dir = {"" = "../.."}
-packages = ["mloda.community.compute_frameworks.example", "mloda.community.extenders.example", "mloda.community.feature_groups.data_operations", "mloda.community.feature_groups.data_operations.aggregation", "mloda.community.feature_groups.data_operations.arithmetic", "mloda.community.feature_groups.data_operations.row_changing", "mloda.community.feature_groups.data_operations.row_changing.resample", "mloda.community.feature_groups.data_operations.row_preserving", "mloda.community.feature_groups.data_operations.row_preserving.binning", "mloda.community.feature_groups.data_operations.row_preserving.datetime", "mloda.community.feature_groups.data_operations.row_preserving.ema", "mloda.community.feature_groups.data_operations.row_preserving.ffill", "mloda.community.feature_groups.data_operations.row_preserving.frame_aggregate", "mloda.community.feature_groups.data_operations.row_preserving.offset", "mloda.community.feature_groups.data_operations.row_preserving.percentile", "mloda.community.feature_groups.data_operations.row_preserving.point_arithmetic", "mloda.community.feature_groups.data_operations.row_preserving.rank", "mloda.community.feature_groups.data_operations.row_preserving.scalar_aggregate", "mloda.community.feature_groups.data_operations.row_preserving.scalar_arithmetic", "mloda.community.feature_groups.data_operations.row_preserving.sessionization", "mloda.community.feature_groups.data_operations.row_preserving.time_bucketization", "mloda.community.feature_groups.data_operations.row_preserving.window_aggregation", "mloda.community.feature_groups.data_operations.string", "mloda.community.feature_groups.example", "mloda.community.feature_groups.example.example_a", "mloda.community.feature_groups.example.example_b"]
+packages = ["mloda.community.compute_frameworks.example", "mloda.community.extenders.example", "mloda.community.feature_groups.data_operations", "mloda.community.feature_groups.data_operations.aggregation", "mloda.community.feature_groups.data_operations.row_changing", "mloda.community.feature_groups.data_operations.row_changing.resample", "mloda.community.feature_groups.data_operations.row_preserving", "mloda.community.feature_groups.data_operations.row_preserving.arithmetic", "mloda.community.feature_groups.data_operations.row_preserving.binning", "mloda.community.feature_groups.data_operations.row_preserving.datetime", "mloda.community.feature_groups.data_operations.row_preserving.ema", "mloda.community.feature_groups.data_operations.row_preserving.ffill", "mloda.community.feature_groups.data_operations.row_preserving.frame_aggregate", "mloda.community.feature_groups.data_operations.row_preserving.offset", "mloda.community.feature_groups.data_operations.row_preserving.percentile", "mloda.community.feature_groups.data_operations.row_preserving.point_arithmetic", "mloda.community.feature_groups.data_operations.row_preserving.rank", "mloda.community.feature_groups.data_operations.row_preserving.scalar_aggregate", "mloda.community.feature_groups.data_operations.row_preserving.scalar_arithmetic", "mloda.community.feature_groups.data_operations.row_preserving.sessionization", "mloda.community.feature_groups.data_operations.row_preserving.time_bucketization", "mloda.community.feature_groups.data_operations.row_preserving.window_aggregation", "mloda.community.feature_groups.data_operations.string", "mloda.community.feature_groups.example", "mloda.community.feature_groups.example.example_a", "mloda.community.feature_groups.example.example_b"]
